Sounds like a good first version
