+++ deepomatic/cli/__init__.py
@@ -1,1 +1,1 @@
-from deepomatic.cli.version import __version__
+from .version import __version__
