+++ deepomatic/cli/cmds/studio_helpers/task.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 import time
 
+
 class Task(object):
     def __init__(self, helper, pk=None, files=None):
         self._helper = helper
@@ -9,7 +10,7 @@ class Task(object):
         ret = self._helper.get('manage/tasks/{}/'.format(task_id))
         if not wait:
             return ret
-        while ret['next'] != None and ret['status'] != 'SUCCESS':
+        while ret['next'] is not None and ret['status'] != 'SUCCESS':
             if ret['status'] in ('FAILURE', 'REVOKED'):
                 raise RuntimeError("Task {} stoped with status".format(task_id))
             elif ret['status'] == 'SUCCESS':
